+++ app/converter/blocks.ts
@@ -309,7 +309,7 @@ const ImageBlock = {
         const alt = decodeHTML(altRaw);
         const url = decodeHTML(urlRaw), attribs = decodeHTML(attribsRaw);
 
-        const img = `<p><img src="${escapeXmlText(url)}" alt="${escapeXmlText(alt)}"` +
+        const img = `<p><${TagNames.Media} src="${escapeXmlText(url)}" alt="${escapeXmlText(alt)}"` +
             `${legalizeAttributes(attribs, ['src', 'alt'], e => console.warn(e))} /></p>`;
 
         if (id !== '') {