+++ app/converter/mftext.ts
@@ -534,7 +534,7 @@ export function expandMfText(node: Element): void {
             const alt = decodeHTML(altRaw);
             const url = decodeHTML(urlRaw), attribs = decodeHTML(attribsRaw);
 
-            return `<img src="${escapeXmlText(url)}" alt="${escapeXmlText(alt)}"` +
+            return `<${TagNames.Media} src="${escapeXmlText(url)}" alt="${escapeXmlText(alt)}"` +
                 `${legalizeAttributes(attribs, ['src', 'alt'], m => console.warn(m))} />`;
         }
     ), isNonVerbatimElement);
@@ -580,7 +580,7 @@ export function expandMfText(node: Element): void {
                         if (match) {
                             let [_1, _2, url, attribs] = match;
                             attribs = legalizeAttributes(attribs || '', ['src', 'alt'], e => console.warn(e));
-                            output.unshift(`<img src="${escapeXmlText(url)}" alt="${escapeXmlText(text)}"${attribs} />`);
+                            output.unshift(`<${TagNames.Media} src="${escapeXmlText(url)}" alt="${escapeXmlText(text)}"${attribs} />`);
                         } else {
                             output.unshift(`<${TagNames.Error}>Failed to parse the link target: <code>` +
                                 escapeXmlText(linkTarget) +